Revert "Revert "Refactor audio""
